backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.8.0
python-dotenv>=1.0.0
httpx>=0.25.0
boto3>=1.34.0
strands-agents>=1.0.0
diagrams>=0.24.0
